+++ modules/lemmaDigitalBidAdapter.js
@@ -1,4 +1,5 @@
 import * as utils from '../src/utils.js';
+import { getDNT } from '../libraries/dnt/index.js';
 import { config } from '../src/config.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { BANNER, VIDEO } from '../src/mediaTypes.js';
@@ -448,7 +449,7 @@ export var spec = {
     var params = request && request.params ? request.params : null;
     if (params) {
       return {
-        dnt: utils.getDNT() ? 1 : 0,
+        dnt: getDNT() ? 1 : 0,
         ua: navigator.userAgent,
         language: (navigator.language || navigator.browserLanguage || navigator.userLanguage || navigator.systemLanguage),
         w: (utils.getWinDimensions().screen.width || utils.getWinDimensions().innerWidth),
