+++ test/spec/modules/mediaforceBidAdapter_spec.js
@@ -1,6 +1,7 @@
 import {assert} from 'chai';
 import {spec, resolveFloor} from 'modules/mediaforceBidAdapter.js';
 import * as utils from '../../../src/utils.js';
+import { getDNT } from 'libraries/dnt/index.js';
 import {BANNER, NATIVE, VIDEO} from '../../../src/mediaTypes.js';
 
 describe('mediaforce bid adapter', function () {
@@ -125,7 +126,7 @@ describe('mediaforce bid adapter', function () {
       ]
     };
 
-    const dnt = utils.getDNT() ? 1 : 0;
+    const dnt = getDNT() ? 1 : 0;
     const secure = window.location.protocol === 'https:' ? 1 : 0;
     const pageUrl = window.location.href;
     const timeout = 1500;
