+++ test/spec/modules/mgidBidAdapter_spec.js
@@ -2,6 +2,7 @@ import {expect} from 'chai';
 import { spec, storage } from 'modules/mgidBidAdapter.js';
 import { version } from 'package.json';
 import * as utils from '../../../src/utils.js';
+import { getDNT } from 'libraries/dnt/index.js';
 import {USERSYNC_DEFAULT_CONFIG} from '../../../src/userSync.js';
 import {config} from '../../../src/config.js';
 
@@ -22,7 +23,7 @@ describe('Mgid bid adapter', function () {
   });
   const screenHeight = screen.height;
   const screenWidth = screen.width;
-  const dnt = utils.getDNT() ? 1 : 0;
+  const dnt = getDNT() ? 1 : 0;
   const language = navigator.language ? 'language' : 'userLanguage';
   let lang = navigator[language].split('-')[0];
   if (lang.length !== 2 && lang.length !== 3) {
