+++ test/spec/modules/openxBidAdapter_spec.js
@@ -4,6 +4,7 @@ import {newBidder} from 'src/adapters/bidderFactory.js';
 import {BANNER, NATIVE, VIDEO} from 'src/mediaTypes.js';
 import {config} from 'src/config.js';
 import * as utils from 'src/utils.js';
+import * as dnt from 'libraries/dnt/index.js';
 // load modules that register ORTB processors
 import 'src/prebid.js'
 import 'modules/currency.js';
@@ -1053,7 +1054,7 @@ describe('OpenxRtbAdapter', function () {
         let doNotTrackStub;
 
         beforeEach(function () {
-          doNotTrackStub = sinon.stub(utils, 'getDNT');
+          doNotTrackStub = sinon.stub(dnt, 'getDNT');
         });
         afterEach(function() {
           doNotTrackStub.restore();
